+++ src/spritecache.h
@@ -54,7 +54,7 @@ void IncreaseSpriteLRU();
 
 void ReadGRFSpriteOffsets(byte container_version);
 size_t GetGRFSpriteOffset(uint32 id);
-bool LoadNextSprite(int load_index, byte file_index, uint file_sprite_id, byte container_version);
+bool LoadNextSprite(int load_index, uint file_index, uint file_sprite_id, byte container_version);
 bool SkipSpriteData(byte type, uint16 num);
 void DupSprite(SpriteID old_spr, SpriteID new_spr);
